Refactor: classroom, reservation 패키지 타입 수정

+++ src/main/java/com/groomiz/billage/classroom/dto/request/AdminClassroomStatusRequest.java
@@ -22,5 +22,5 @@ public class AdminClassroomStatusRequest {
 	private List<Long> buildings;
 
 	@Schema(description = "층 리스트", example = "[1, 2, 3]")
-	private List<Integer> floors;
+	private List<Long> floors;
 }
